gitlab issue #54
So, after a couple of hours of looking online, it seems that to access
or upload any json file, we need to store it somewhere(temoprarily
atleast). Therefore I used session storage to store our temorary json
file. I made a button on panel.html then takes in a file input(user is
restricted to only json files) and then puts that file in session
storage.