Fixes bug created by removing Upload Template
After removing the Upload Template button from the Template Builder page,
the Inconsistency Checker page stopped working (doing comparisons).
This was caused because there was an event listener for the Upload Template
button on the Template Builder page, which was confused when we removed the button.

This commit fixes the bug, and the Inconsistency Checker works as normal again.